+++ types.ts
@@ -2,13 +2,14 @@ import type { Engine, Helper } from "./core/renderer.ts";
 import type { Data as PageData, Page } from "./core/file.ts";
 import type { Loader } from "./core/fs.ts";
 import type { default as Site, Plugin } from "./core/site.ts";
-import type { Archetype } from "./cli/create.ts";
+import type { Archetype, ArchetypeFile } from "./cli/create.ts";
 import type { Middleware, RequestHandler } from "./core/server.ts";
 
 declare global {
   namespace Lume {
     export type {
       Archetype,
+      ArchetypeFile,
       Engine,
       Loader,
       Middleware,