+++ CHANGELOG.md
@@ -10,6 +10,8 @@ Go to the `v1` branch to see the changelog of Lume 1.
 ### New
 - Predefined archetypes for common code scaffolds
   - `plugin` to create a Lume plugin in the local folder `_plugins/[name].ts`.
+  - `archetype` to create a Lume archetype in the local folder `_archetype/[name].ts`.
+  - `cms` to create the `_cms.ts` file.
 
 ### Changed
 - The `merge` utility function changed the types to better reflect its returned type.
@@ -26,6 +28,7 @@ Go to the `v1` branch to see the changelog of Lume 1.
   const options = merge(defaults, userOptions);
   ```
 - `favicon` plugin: Don't override favicons added explicitly with `site.add()` or `site.copy()`.
+- Archetypes are loaded from the cwd, not the `src` folder.
 
 ### Fixed
 - `icons` plugin: Added missing type for `spriteIcon` helper [#852].